Clearing ephemeral RDS data each time, before processing.
This library preserves some values (PS, PTYN, etc.), but other
values are considered "ephemeral", and are cleared each time
new RDS data is received before processing.

It is the responsibility of the host to save any values when
notified of new RDS data. This change is so that the host can
know when values are no longer present - or change frequently.